refactor: Improved policy, info and bug icons and ref refactor: Renamed Help page to About and ref
